Add cpu cycle counting
Instead of assigning each table entry a cycle count from the refernce
manual, this commit tries to find a account for cycles
by specifying the cost to Read/Write or Modify operands for  each Address
Mode.

It would probably be a lot easier to just specify each instructions
cycle  count in the lookup table.