No more emulator panics
Did the big mistake of not worrying about error propagation
and now had to do a long and tedious refactoring to wrap everything
in Result<>s.

Added a test to validate that any random permutation of bytes as
an operation is not going to cause a panic.

This will allow us to halt the emulator on errors to investigate in the
debug view of RES.