Add foundation for audio output
Using cpal to play audio buffers. And using fundsp to generate
samples.

To get cpal to work in the browser, we have to initialize audio
during a user interaction. So we have to start the emulator after
a click.

Unfortunately, a click in egui will not be sufficient since it's
not handled via a JS event handler, but during animation frames.

For now just plays noise.