Add gamepad support using gilrs
Mappings are a pain on osx. None of my gamepads actually have
proper SDL or Web gamepad mappings.
The only one that works somewhat consistently is the PS5 controller.

Either way, if the gamepad is properly mapped, the emulator can now use
it and should work on all platforms thanks to gilrs.