res_egui: Sync to wall time and reduce audio cracking
egui does not provide a consistent framerate that would allow us
to maintain a consistent stream of audio samples.
To get around this, the emulator is synchronized to wall time and
frames are skipped, or multiple executed as needed.
This provides a more consistent stream of audio samples and
does not crack during normal operation.
Switching applications, or overlays can cause enough jitter in the
frame rate that audio might still crack.
(Overlays such as the volume control on OSX cause a lot of jitter..
probably since the application is switched between HW overlays and
compositing)